+++ src/cmd/ffmpeg_cmd/flag.go
@@ -1,19 +1,7 @@
-package cmd
+package ffmpeg_cmd
 
 import (
 	"github.com/urfave/cli/v2"
-)
-
-var (
-	proxyFlag   = &cli.StringFlag{Name: "proxy", Usage: "代理配置,如: http://127.0.0.1:1080"}
-	destDirFlag = &cli.StringFlag{Name: "dest_dir", Usage: "设置下载目录", Value: "./javbus"}
-	siteFlag    = &cli.StringFlag{
-		Name:        "site",
-		Usage:       "选择爬取数据的网站,支持网站(javbus)",
-		Destination: nil,
-		HasBeenSet:  false,
-		Value:       "javbus",
-	}
 )
 
 var (
@@ -61,11 +49,6 @@ var (
 		Name:  "exec",
 		Value: false,
 		Usage: "是否执行批处理命令False时仅打印命令",
-	}
-	fonts_flag = &cli.StringSliceFlag{
-		Name:    "fonts",
-		Aliases: []string{"f"},
-		Usage:   "添加字体文件",
 	}
 	fonts_path_flag = &cli.StringFlag{
 		Name:  "fonts_path",