+++ src/cmd/jav_cmd/code.go
@@ -1,14 +1,14 @@
-package cmd
+package jav_cmd
 
 import (
 	"path"
 
-	"github.com/gsxhnd/garage/crawl"
-	"github.com/gsxhnd/garage/utils"
+	"github.com/gsxhnd/garage/src/jav"
+	"github.com/gsxhnd/garage/src/utils"
 	"github.com/urfave/cli/v2"
 )
 
-var javCodeCmd = &cli.Command{
+var CodeCmd = &cli.Command{
 	Name:      "jav_code",
 	Aliases:   nil,
 	Usage:     "根据指定番号爬取数据",
@@ -23,7 +23,7 @@ var javCodeCmd = &cli.Command{
 			logger  = utils.GetLogger()
 			code    = ctx.Args().Get(0)
 			proxy   = ctx.String("proxy")
-			destDir = "./javs"
+			destDir = ctx.String("dest_dir")
 		)
 
 		if err := utils.MkdirDestDir(path.Join(destDir, code)); err != nil {
@@ -31,8 +31,13 @@ var javCodeCmd = &cli.Command{
 			return err
 		}
 
-		c := crawl.NewCrawlClient(logger)
-		_ = c.SetProxy(proxy)
+		c := jav.NewCrawlClient(logger)
+		err := c.SetProxy(proxy)
+		if err != nil {
+			logger.Panic("crawl set proxy error: " + err.Error())
+			return err
+		}
+
 		c.StarCrawlJavbusMovie(code)
 		return nil
 	},